fogl-102 Move coap to device_api/coap. Create consistent starter modules with start method
