+++ src/python/foglamp/model/starter.py
@@ -1,0 +1,6 @@
+from .config import read_config
+
+
+def start():
+    read_config()
+
